requirements.txt
fastapi==0.116.1
hezar==0.42.0
numpy==2.3.2
pandas==2.3.1
Pillow==11.3.0
reportlab==4.4.3
Requests==2.32.4
scikit_learn==1.7.1
streamlit==1.48.0
torch==2.8.0
tqdm==4.67.1
transformers==4.55.0
ultralytics==8.3.176
uvicorn==0.35.0
python-multipart==0.0.20
python-dotenv==1.1.1
uuid==1.30